requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.5
djangorestframework==3.15.2
pymongo==3.11.4
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
mysqlclient>=2.1.0
mongoengine>=0.29.1
django-cors-headers>=4.1.0
Pillow>=10.0.0